requirements.txt
# pip freeze > requirements.txt:

blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
PyJWT==2.8.0
python-decouple==3.8
Werkzeug==3.0.0

# flask
# flask-mysqldb
# PyJWT
# python-decouple
